+++ packages/create-airbnb-x-config/utils/strings/index.ts
@@ -1,0 +1,1 @@
+export * from '@/utils/strings/paths';
