+++ packages/create-airbnb-x-config/utils/strings/paths.ts
@@ -1,0 +1,3 @@
+import path from 'node:path';
+
+export const rootPath = path.resolve('.');
